requirements.txt
coverage
cryptography
docutils
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT==0.3.2
Flask-JWT-Extended==3.17.0
PyJWT==1.4.2
pylint
pytest
pytest-coverage
python-dateutil
python-dotenv
requests
urllib3
behave
testing.postgresql
sqlalchemy
postgres